Moved the thread for console input to main function
In the ready() function when application commands are registered this
blocks the event listener for some reason. However, since we only get a
Context (which is needed for several functions called from the console
listener) in the ready function we have to create a channel and put a
Sender into an Arc<Mutex<>> inside the Handler so it can be passed to
the function savely.